requirements.txt
# ================================
# HTTP 통신 (KIS API 호출)
# ================================
requests

# ================================
# 환경 변수 로드 (.env)
# ================================
python-dotenv

# ================================
# MariaDB / MySQL 연결
# ================================
pymysql

